requirements.txt
pytubefix
rich
ffmpeg-python